Add reference to vertices (#160)
In graph theory words nodes and vertices are synonyms, as are edges and links. In my experience network theory, which is a subset of graph theory, more commonly uses nodes. Where a network is a graph where nodes or edges possess attributes.

(I should add this to the answer to reviewers comments...)